configure: allow cross-compilation
Some compilers are in fact a wrapper (e.g. to speed compilation with
ccache or distcc, or as a cross-compiler wrapper). Those wrappers may
not properly recognise the -v option, or may internally enforce some
flags that conflict with -v.

Use --version as a fall-back.

With --version. the compiler will report its executable's basename, e.g.
arm-linux-gcc not gcc. Catter for that by grepping for the compiler
family name.

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
[yann.morin.1998@free.fr:
  - retain "$CC -v" as default, fallback to "$CC --version"
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>