"caches" successful hash algorithms for a given checkfile
The goal, based on a sort of principle of locality (or just the
typical way checkfiles are generated), is to improve real-world
performance by inferring the correct hash algorithm based on
previous successes in the same run.  I imagine in most cases
there will be exactly one successful algorithm, but this
implementation supports arbitrarily many.

Signed-off-by: Shawn Bayern <sbayern@law.fsu.edu>